requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
eventlet==0.35.2
python-dotenv==1.0.1
Werkzeug==3.0.3
